added proper css and values for the comparison table
